# backend-java/model/src/main/resources/admin/openapi-admin.yaml
openapi: 3.0.3
info:
  title: Wust Dormitory Select 管理端接口
  version: 1.0.0
paths:
  /api/v1/admin/dashboard:
    get:
      tags: [Admin]
      operationId: getAdminDashboard
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/majors:
    get:
      tags: [Admin]
      operationId: listMajors
      security: [{ bearerAuth: [] }]
      parameters:
        - name: enabled
          in: query
          required: false
          schema: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }
    post:
      tags: [Admin]
      operationId: createMajor
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MajorRequest' }
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/majors/{id}:
    put:
      tags: [Admin]
      operationId: updateMajor
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MajorRequest' }
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }
        '404': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/students:
    get:
      tags: [Admin]
      operationId: listStudents
      security: [{ bearerAuth: [] }]
      parameters:
        - name: keyword
          in: query
          schema: { type: string }
        - name: gender
          in: query
          schema:
            type: string
            pattern: '^[MF]$'
        - name: majorId
          in: query
          schema: { type: integer, format: int64 }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }
    post:
      tags: [Admin]
      operationId: createStudent
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StudentRequest' }
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/students/{id}:
    put:
      tags: [Admin]
      operationId: updateStudent
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StudentRequest' }
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }
        '404': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/students/import:
    post:
      tags: [Admin]
      operationId: importStudents
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 5000
              items: { $ref: '#/components/schemas/StudentRequest' }
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }
        '400': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/buildings:
    get:
      tags: [Admin]
      operationId: listBuildings
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }

  /api/v1/admin/rooms:
    get:
      tags: [Admin]
      operationId: listRooms
      security: [{ bearerAuth: [] }]
      parameters:
        - name: buildingId
          in: query
          schema: { type: integer, format: int64 }
        - name: gender
          in: query
          schema:
            type: string
            pattern: '^[MF]$'
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }

  /api/v1/admin/rooms/{roomId}:
    put:
      tags: [Admin]
      operationId: updateRoom
      security: [{ bearerAuth: [] }]
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoomRequest' }
      responses:
        '200': { $ref: '#/components/responses/VoidSuccess' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/batches:
    get:
      tags: [Admin]
      operationId: listBatches
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }
    post:
      tags: [Admin]
      operationId: createBatch
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchRequest' }
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }
        '400': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/batches/{batchId}/prepare:
    post:
      tags: [Admin]
      operationId: prepareBatch
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BatchIdPath'
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }

  /api/v1/admin/batches/{batchId}/status/{targetStatus}:
    post:
      tags: [Admin]
      operationId: changeBatchStatus
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BatchIdPath'
        - name: targetStatus
          in: path
          required: true
          schema:
            type: string
            pattern: '^(PUBLISHED|OPEN|PAUSED|CLOSED|ALLOCATING|FINISHED|CANCELLED)$'
      responses:
        '200': { $ref: '#/components/responses/VoidSuccess' }
        '400': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/batches/{batchId}/allocation/preview:
    get:
      tags: [Admin]
      operationId: previewAllocation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BatchIdPath'
        - name: randomSeed
          in: query
          required: false
          schema: { type: integer, format: int64, default: 20260801 }
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }

  /api/v1/admin/batches/{batchId}/allocation/commit:
    post:
      tags: [Admin]
      operationId: commitAllocation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BatchIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllocationCommitRequest' }
      responses:
        '200': { $ref: '#/components/responses/ObjectSuccess' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }

  /api/v1/admin/audit-logs:
    get:
      tags: [Admin]
      operationId: listAuditLogs
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }

  /api/v1/admin/batches/{batchId}/assignments.csv:
    get:
      tags: [Admin]
      operationId: exportAssignments
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BatchIdPath'
      responses:
        '200':
          description: 分配结果CSV文件
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            text/csv:
              schema: { type: string, format: binary }

components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    BatchIdPath:
      name: batchId
      in: path
      required: true
      schema: { type: integer, format: int64 }

  responses:
    ObjectSuccess:
      description: 操作成功
      content:
        application/json:
          schema: { $ref: '../common-response/openapi-common-response.yaml#/components/schemas/ObjectSuccessResponse' }
    ListSuccess:
      description: 查询成功
      content:
        application/json:
          schema: { $ref: '../common-response/openapi-common-response.yaml#/components/schemas/ListSuccessResponse' }
    VoidSuccess:
      description: 操作成功
      content:
        application/json:
          schema: { $ref: '../common-response/openapi-common-response.yaml#/components/schemas/VoidSuccessResponse' }
    ErrorResponse:
      description: 业务错误
      content:
        application/json:
          schema: { $ref: '../common-response/openapi-common-response.yaml#/components/schemas/ErrorResponse' }

  schemas:
    MajorRequest:
      type: object
      required: [majorCode, majorName, enabled]
      properties:
        majorCode: { type: string, minLength: 1, maxLength: 32 }
        majorName: { type: string, minLength: 1, maxLength: 128 }
        enabled: { type: boolean }

    StudentRequest:
      type: object
      required: [studentNumber, studentName, gender, majorId]
      properties:
        studentNumber: { type: string, pattern: '^\d{12}$' }
        studentName: { type: string, minLength: 1, maxLength: 128 }
        gender: { type: string, pattern: '^[MF]$' }
        majorId: { type: integer, format: int64, minimum: 1 }

    RoomRequest:
      type: object
      required: [roomType, capacity, gender, operationalStatus, reason]
      properties:
        roomType:
          type: string
          pattern: '^(FOUR_PERSON|FIVE_PERSON|SIX_PERSON|OTHER)$'
        capacity: { type: integer, minimum: 1, maximum: 20 }
        gender: { type: string, pattern: '^[MF]$' }
        operationalStatus:
          type: string
          pattern: '^(ENABLED|DISABLED|MAINTENANCE)$'
        remark: { type: string, maxLength: 500 }
        reason: { type: string, minLength: 1, maxLength: 500 }

    BatchRequest:
      type: object
      required: [batchCode, batchName, startAt, endAt, holdDurationSeconds, allowTeam, teamMaxSize, allowStudentRandom]
      properties:
        batchCode: { type: string, minLength: 1, maxLength: 32 }
        batchName: { type: string, minLength: 1, maxLength: 128 }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        holdDurationSeconds: { type: integer, minimum: 30, maximum: 3600 }
        allowTeam: { type: boolean }
        teamMaxSize: { type: integer, minimum: 2, maximum: 20 }
        allowStudentRandom: { type: boolean }

    AllocationCommitRequest:
      type: object
      required: [randomSeed, idempotencyKey]
      properties:
        randomSeed: { type: integer, format: int64 }
        idempotencyKey: { type: string, minLength: 1, maxLength: 128 }

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token